backend/pyproject.toml
[project]
name = "ianoie"
version = "0.1.0"
description = "GPU AI App Platform for NVIDIA DGX"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "docker[tls]>=7.1",
    "pyyaml>=6.0",
    "celery[redis]>=5.4",
    "redis>=5.2",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "pynvml>=12.0",
    "alembic>=1.14",
    "structlog>=24.4",
    "httpx>=0.28",
    "cryptography>=43.0",
    "psycopg2-binary>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.8",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
